+++ js/util.js
@@ -8,4 +8,4 @@ const getRand = (min, max) => {
   return Math.random() * (max - min + 1)  + min;
 };
 
-postPhotos();
+export {getRand};
